refactor: Split types/lib.rs into focused modules
Refactored the 620-line crates/types/src/lib.rs into a logical module
structure for better maintainability and navigation:

**Module structure:**
- lib.rs (16 lines) - Module declarations and re-exports
- data_type.rs (89 lines) - DataType enum and compatibility logic
- temporal/ - Date/time types (316 lines total)
  - mod.rs - IntervalField enum and re-exports
  - date.rs - Date type with parsing and comparison
  - time.rs - Time type with fractional seconds
  - timestamp.rs - Timestamp combining date and time
  - interval.rs - Interval duration type
- sql_value/ - Runtime value representation (269 lines total)
  - mod.rs - SqlValue enum and basic methods
  - comparison.rs - PartialOrd and Eq implementations
  - hash.rs - Hash with NaN handling
  - display.rs - Display formatting

**Benefits:**
- Each file under ~100 lines for easier comprehension
- Related types grouped logically (temporal, value operations)
- Trait implementations separated by concern
- No public API changes (all types re-exported from lib.rs)

**Testing:**
- All 91 tests pass in types crate
- Dependent crates (executor, parser, catalog) compile successfully
- No import changes needed in consuming code

Closes #1147
