+++ crates/types/src/temporal/mod.rs
@@ -1,0 +1,30 @@
+//! Temporal types for SQL:1999
+//!
+//! This module contains all date/time related types including:
+//! - DATE
+//! - TIME
+//! - TIMESTAMP
+//! - INTERVAL
+
+mod date;
+mod time;
+mod timestamp;
+mod interval;
+
+pub use date::Date;
+pub use time::Time;
+pub use timestamp::Timestamp;
+pub use interval::Interval;
+
+/// SQL:1999 Interval Fields
+///
+/// Represents the fields that can be used in INTERVAL types
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum IntervalField {
+    Year,
+    Month,
+    Day,
+    Hour,
+    Minute,
+    Second,
+}
